fix: 修改 react-native-vision-camera的测试demo
