+++ react-native-vision-camera/test/Photo/FormatTest.tsx
@@ -1,0 +1,66 @@
+import {Button, StyleSheet, Text, View} from 'react-native';
+import React, {useRef, useState} from 'react';
+import {
+  Camera,
+  useCameraDevice,
+  useCameraFormat,
+  useCameraPermission,
+} from 'react-native-vision-camera';
+import {TestSuite, TestCase, Tester} from '@rnoh/testerino';
+
+export function FormatTest() {
+  const device = useCameraDevice('back');
+  const format = useCameraFormat(device, [
+    {videoResolution: {width: 1920, height: 1080}},
+    {fps: 30},
+  ]);
+  const {hasPermission, requestPermission} = useCameraPermission();
+  const camera = useRef<Camera>(null);
+
+  if (!device) {
+    return <Text>No Devices</Text>;
+  }
+
+  if (!hasPermission) {
+    requestPermission();
+  }
+
+  return (
+    <Tester>
+      <TestSuite name="format">
+        <TestCase itShould={``}>
+          <View>
+            <Text style={style.text}>format:{JSON.stringify(format)}</Text>
+          </View>
+          <Camera
+            style={style.cameraPreview}
+            ref={camera}
+            device={device}
+            isActive
+            preview
+            photo
+            format={format}
+          />
+        </TestCase>
+      </TestSuite>
+    </Tester>
+  );
+}
+
+const style = StyleSheet.create({
+  cameraPreview: {width: 300, height: 400},
+  actionBtn: {
+    flexDirection: 'row',
+    flexWrap: 'wrap',
+    justifyContent: 'space-evenly',
+    padding: 10,
+    gap: 5,
+    position: 'absolute',
+    top: 300,
+  },
+  text: {
+    // fontSize: 20,
+    // textAlign: 'center',
+    // color: '#fff',
+  },
+});
