+++ react-native-vision-camera/test/Video/FormatTest.tsx
@@ -3,14 +3,16 @@ import React, {useRef, useState} from 'react';
 import {
   Camera,
   useCameraDevice,
-  useCameraFormat,
   useCameraPermission,
+  useCameraDevices,
+  useCameraFormat,
 } from 'react-native-vision-camera';
 import {TestSuite, TestCase, Tester} from '@rnoh/testerino';
 import {CameraRoll} from '@react-native-camera-roll/camera-roll';
 
-export function AudioTest() {
+export function FormatTest() {
   const device = useCameraDevice('back');
+  const devices = useCameraDevices();
   const {hasPermission, requestPermission} = useCameraPermission();
   const camera = useRef<Camera>(null);
   const format = useCameraFormat(device, [
@@ -30,12 +32,15 @@ export function AudioTest() {
   const [flash, setFlash] = useState<'off' | 'on'>('off');
   const [preview, setPreview] = useState(true);
   const [videoCodec, setVideoCodec] = useState<'h265' | 'h264'>('h265');
-
+  const [videoStabilizationMode, setVideoStabilizationMode] = useState<
+    'off' | 'standard' | 'cinematic' | 'cinematic-extended' | 'auto'
+  >('auto');
   const [videoHdr, setVideoHdr] = useState(false);
 
   const [startStatus, seteStartStatus] = useState('end');
   const [videoFile, setVideoFile] = useState<string>('');
   const [videoPath, setVideoPath] = useState<string>('');
+  const [fps, set] = useState<number>(30);
 
   const onStart = async () => {
     setVideoPath('');
@@ -73,37 +78,14 @@ export function AudioTest() {
     seteStartStatus('start');
     camera.current?.resumeRecording();
   };
-  const onChangeVideoCodeC = async () => {
-    const code = videoCodec === 'h264' ? 'h265' : 'h264';
-    setVideoCodec(code);
-  };
-
-  const onChangeAudio = async () => {
-    setAudio(!audio);
-  };
 
   return (
     <Tester>
-      <TestSuite name="audio">
-        <TestCase itShould={`启用视频录制的音频捕获:${audio ? '是' : '否'}`}>
+      <TestSuite name="format">
+        <TestCase itShould={``}>
           <View>
-            <Text>录像结果:{videoFile}</Text>
+            <Text>format:{JSON.stringify(format)}</Text>
           </View>
-          {videoPath && (
-            <View style={{height: 50}}>
-              <Button
-                title="SaveAsset"
-                onPress={() => {
-                  CameraRoll.saveAsset(videoPath).then(res => {
-                    setTimeout(() => {
-                      setVideoPath('');
-                      setVideoFile('');
-                    }, 500);
-                  });
-                }}
-              />
-            </View>
-          )}
           <Camera
             style={style.cameraPreview}
             ref={camera}
@@ -111,47 +93,12 @@ export function AudioTest() {
             preview={preview}
             device={device}
             video={true}
-            audio={audio}
+            audio={false}
             videoHdr={videoHdr}
+            videoStabilizationMode={videoStabilizationMode}
             fps={30}
             format={format}
           />
-
-          <View style={style.actionBtn}>
-            <>
-              {videoHdr && videoCodec === 'h264' ? (
-                <Text>videoHdr为true时，videoCodeC只能设置为h265; </Text>
-              ) : (
-                <>
-                  {startStatus === 'end' ? (
-                    <Button title="开始" onPress={onStart}></Button>
-                  ) : (
-                    ''
-                  )}
-                  {startStatus === 'start' ? (
-                    <Button title="暂停" onPress={onPause}></Button>
-                  ) : (
-                    ''
-                  )}
-                  {startStatus === 'pause' ? (
-                    <Button title="恢复" onPress={onResume}></Button>
-                  ) : (
-                    ''
-                  )}
-                  {startStatus !== 'end' ? (
-                    <Button title="停止" onPress={onStop}></Button>
-                  ) : (
-                    ''
-                  )}
-                </>
-              )}
-              <Text>fps: 30</Text>
-            </>
-            <Button
-              title={`codec:${videoCodec}`}
-              onPress={onChangeVideoCodeC}></Button>
-            <Button title={`audio:${audio}`} onPress={onChangeAudio}></Button>
-          </View>
         </TestCase>
       </TestSuite>
     </Tester>
@@ -159,18 +106,16 @@ export function AudioTest() {
 }
 
 const style = StyleSheet.create({
+  cameraPreview: {
+    width: 300,
+    height: 400,
+  },
   actionBtn: {
     flexDirection: 'row',
     flexWrap: 'wrap',
     justifyContent: 'space-between',
     padding: 10,
-    gap: 10,
-    position: 'absolute',
-    top: 300,
-  },
-  cameraPreview: {
-    width: 300,
-    height: 400,
+    gap: 5,
   },
   text: {
     fontSize: 20,
